pyproject.toml
[tool.poetry]
name = "easy-cv"
version = "0.1.0"
description = ""
authors = ["adam <adam.hoffstein@gmail.com>"]
readme = "README.md"
packages = [{include = "easy_cv"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
django-simple-history = "^3.3.0"
django-model-utils = "^4.3.1"
django-crispy-forms = "^2.0"
crispy-bootstrap5 = "^0.7"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
